Fixing Issue #237 - Alternate Uptime Instance
Spine will not look at non-timetics uptime when sysUpTimeInstance overflows

+++ spine.h
@@ -525,7 +525,7 @@ typedef struct host_struct {
 	int    snmp_retries;
 	char   snmp_sysDescr[600];
 	char   snmp_sysObjectID[160];
-	u_int  snmp_sysUpTimeInstance;
+	unsigned long long snmp_sysUpTimeInstance;
 	char   snmp_sysContact[300];
 	char   snmp_sysName[300];
 	char   snmp_sysLocation[600];
